+++ include/axl/axl-version.h
@@ -26,8 +26,8 @@
 
 #define AXL_VERSION_MAJOR   0
 #define AXL_VERSION_MINOR   19
-#define AXL_VERSION_PATCH   1
-#define AXL_VERSION_STRING  "0.19.1"
+#define AXL_VERSION_PATCH   2
+#define AXL_VERSION_STRING  "0.19.2"
 
 /* Encoded as 0xMMmmpp (major, minor, patch) for simple comparisons. */
 #define AXL_VERSION_NUMBER \